+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Chord.getPitchByDegree
 
+- Chord.transpose & Scale.getTransposedScale
+
 ### Changed
 
 - Renamed Scale.degree to Scale.getPitchByDegree